requirements.txt
streamlit
pandas
langchain
langchain-groq
python-dotenv
plotly